+++ app/src/main/java/io/github/muntashirakon/AppManager/permissions/PermissionAppsAdapter.java
@@ -8,6 +8,7 @@ import android.view.ViewGroup;
 import android.widget.ImageView;
 
 import androidx.annotation.NonNull;
+import androidx.appcompat.widget.TooltipCompat;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.google.android.material.button.MaterialButton;
@@ -87,7 +88,7 @@ class PermissionAppsAdapter extends RecyclerView.Adapter<PermissionAppsAdapter.V
                 ? R.string.permission_reference_restore
                 : R.string.permission_reference_pin;
         h.referenceAction.setContentDescription(h.itemView.getResources().getString(referenceDescription));
-        h.referenceAction.setTooltipText(h.itemView.getResources().getString(referenceDescription));
+        TooltipCompat.setTooltipText(h.referenceAction, h.itemView.getResources().getString(referenceDescription));
         h.referenceAction.setIconResource(r.referenceDrifted ? R.drawable.ic_restore : R.drawable.ic_flag);
         h.referenceAction.setEnabled(mInteractionsEnabled && (!r.referenceDrifted || r.anyModifiable));
         h.referenceAction.setOnClickListener(v -> {